Added fixes

+++ activity/src/main/java/com/kairos/service/shift/ShiftSickService.java
@@ -222,9 +222,11 @@ public class ShiftSickService extends MongoBaseService {
 
         });
         shiftService.createShifts(activity.getUnitId(),ObjectMapperUtils.copyCollectionPropertiesByMapper(allShiftsToUpdate,ShiftDTO.class),ShiftActionType.SAVE);
-        StaffAdditionalInfoDTO staffAdditionalInfoDTO = userIntegrationService.verifyUnitEmploymentOfStaff(DateUtils.asLocalDate(allShiftsToDelete.get(0).getStartDate()), allShiftsToDelete.get(0).getStaffId(), allShiftsToDelete.get(0).getEmploymentId(), Collections.emptySet());
-        shiftService.deleteShifts(new ArrayList<>(),allShiftsToDelete,staffAdditionalInfoDTO);
-        shiftMongoRepository.saveEntities(allShiftsToDelete);
+        if(isCollectionNotEmpty(allShiftsToDelete)) {
+            StaffAdditionalInfoDTO staffAdditionalInfoDTO = userIntegrationService.verifyUnitEmploymentOfStaff(DateUtils.asLocalDate(allShiftsToDelete.get(0).getStartDate()), allShiftsToDelete.get(0).getStaffId(), allShiftsToDelete.get(0).getEmploymentId(), Collections.emptySet());
+            shiftService.deleteShifts(new ArrayList<>(), allShiftsToDelete, staffAdditionalInfoDTO);
+            shiftMongoRepository.saveEntities(allShiftsToDelete);
+        }
 
     }
